// src/app/model/producer/producer.ts
export class Producer {
  id?: string;
  created?: any;
  nome?: string;
  email?: string;
  password?: string;
  fiscalid?: string;
  products?: string;
  province?: string;
  city?: string;
  postalCode?: number;
  address?: string;
  phone?: string;
  description?: string;
  status?: string;
  coverPhoto?: any;
  linkFacebook?: string;
  linkInstagram?: string;
  linkWebsite?: string;
  urlPhoto?: any;


}